requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
httpx
beautifulsoup4
openai
google-genai
google-api-python-client
google-auth
pytest
pytest-cov
